+++ resources/leiningen/new/luminus/reagent/src/cljs/core.cljs
@@ -31,7 +31,7 @@
        [:span][:span][:span]]]
      [:div#nav-menu.navbar-menu
       {:class (when @expanded? :is-active)}
-      [:div.navbar-end
+      [:div.navbar-start
        [nav-link "#/" "Home" :home]
        [nav-link "#/about" "About" :about]]]]))